requirements.txt
pyyaml~=5.4.1
podcastparser~=0.6.7
tqdm~=4.62.1
certifi~=2020.6.20
pymediainfo~=5.1.0
requests~=2.24.0
ffmpeg-python==0.2.0
pyyaml
podcastparser
python-box